upgrade : news detail view
